+++ sqlx-exasol-impl/src/connection/etl/export/mod.rs
@@ -48,13 +48,6 @@ impl AsyncRead for ExaExport {
     }
 }
 
-/// The EXPORT source type, which can either directly be a table or an entire query.
-#[derive(Clone, Copy, Debug)]
-pub enum ExportSource<'a> {
-    Query(&'a str),
-    Table(&'a str),
-}
-
 pub enum ExaExportState {
     /// The worker is fully connected and ready for I/O.
     Poll(ExaExportReader),