+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- [#42](https://github.com/bobozaur/sqlx-exasol/pull/42): Added support for specifying schema in ETL export
 - [#35](https://github.com/bobozaur/sqlx-exasol/pull/35): Compile-time query support
   - Compile-time query validation support.
   - `ExaHasArrayType` marker trait to allow array-like parameter binding for custom types.
@@ -38,6 +39,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Removed
 
+- [#42](https://github.com/bobozaur/sqlx-exasol/pull/42): made ExportSource private
 - [#35](https://github.com/bobozaur/sqlx-exasol/pull/35): Compile-time query support
   - `Encode` implementation for `&mut [T]`.
   - `ExaDataType::Null`.